# apps/yt-transcribe/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "yt-transcribe Worker API",
    "description": "Cloudflare Worker API for downloading, transcribing, and storing YouTube video transcripts.\n\nAll `/admin/*` routes require `Authorization: Bearer <YT_TRANSCRIBE_ADMIN_TOKEN>`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://yt-transcribe-worker.njmalaza.workers.dev",
      "description": "Cloudflare Workers (production)"
    }
  ],
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin token set via `YT_TRANSCRIBE_ADMIN_TOKEN` wrangler secret."
      }
    },
    "schemas": {
      "TranscribeRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://www.youtube.com/watch?v=rdWZo5PD9Ek"
          }
        }
      },
      "TranscribeResponse": {
        "type": "object",
        "properties": {
          "transcript_url": {
            "type": "string",
            "format": "uri",
            "description": "Blob URL where the transcript text file was uploaded."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/transcribe": {
      "post": {
        "summary": "Transcribe a YouTube video",
        "description": "Downloads audio from the given YouTube URL, transcribes it with whisper.cpp, uploads the transcript to blob storage, and returns the blob URL.",
        "operationId": "transcribeVideo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transcription successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing or invalid URL).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Transcription or upload failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service initialisation failed (missing env vars or Infisical error).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/retry-state": {
      "get": {
        "summary": "Get retry state for item",
        "description": "Starts the API container and returns retry metadata from Cloudflare D1 for a specific media item id.",
        "operationId": "getRetryState",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "media_items.id UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retry state payload (or found=false when no row exists).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing id query param.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      }
    }
  }
}